+++ src/engine/components/numericLineGenerator/index.js
@@ -6,13 +6,15 @@
  */
 import { Container } from 'pixi.js';
 
-import { type NumberActor } from 'engine/components/numberGenerator';
+import { type NumberActor, type BeHappyState } from 'engine/components/numberGenerator';
 
 import arrowGenerator from './arrowGenerator';
 import lineGenerator, { type Line } from './lineGenerator';
 
-const receiveNumberAtPositionConfig = (line: Line) => (number: NumberActor, position: number): void => {
-  line.receiveNumberAtPosition(number, position);
+const receiveNumberAtPositionConfig = (line: Line) =>
+  (number: NumberActor, position: number): Promise<void> => line.receiveNumberAtPosition(number, position);
+const beHappyConfig = (line: Line) => (state: BeHappyState): void => {
+  line.beHappy(state);
 };
 
 export type NumericLineData = {|
@@ -21,7 +23,8 @@ export type NumericLineData = {|
 |};
 export type NumericLine = {|
   view: Container,
-  receiveNumberAtPosition(number: NumberActor, position: number): void,
+  receiveNumberAtPosition(number: NumberActor, position: number): Promise<void>,
+  beHappy: (state: BeHappyState) => void,
 |};
 const numericLineGenerator = (numbers: Array<number | null>, size: number, margin: number): NumericLine => {
   const view = new Container();
@@ -39,6 +42,7 @@ const numericLineGenerator = (numbers: Array<number | null>, size: number, margi
   return {
     view,
     receiveNumberAtPosition: receiveNumberAtPositionConfig(line),
+    beHappy: beHappyConfig(line),
   };
 };