+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## HEAD
 
+- 🚀 [2017-05-11] `engine/components/numericLineGenerator`: update to v3.1.0.
+  - 🐛 Fix `receiveNumberAtPosition`: it wasn't returning the promise.
+  - 🚀 New method: `beHappy`. Updates the happy state for each number contained by the `Line`.
+  - 🚀 `./lineGenerator`: simpler generator function.
+  - 🚀 New stories.
 - 🚀 [2017-05-11] `engine/containers/numberGenerator`: update to v3.1.0.
   - 🚀 Add new method `beHappy`: it receives a `state` (`'start' | 'stop'`). Is used when the game has finish correctly.
 - 🚀 [2017-05-11] `engine/helpers/customEases`: v1.0.0!